Handle errors more gracefully
The original version used wrong argument syntax in the crontab example
and also did not handle errors gracefully. A combination which lead to
failures without any hint what’s going on. The error output from OpenSSL
is captured and both the failing command and its error output are
printed if the certificate could not be checked.