Correctly specify server name for SNI
The previous version did not handle SNI correctly as it did not supply
the necessary parameter to OpenSSL to be used by the TLS client. This
lead to wrong reports for web servers hosting multiple domains secured
with SSL. The web server always presented the certificate for the
default domain for all requests by ssl-cert-check.